Skip delete when posted from Telegram User

+++ src/entry.py
@@ -8,5 +8,8 @@ def entry(bot, update):
     # print("\n")
     if update.message:
         if not update.message.reply_to_message:
-            logging.info("DELETING...")
-            bot.delete_message(chat_id=update.message.chat.id, message_id=update.message.message_id)
+            if update.message.from_user.id == 777000:
+                print("Linked Channel's post. Skip delete!")
+            else:
+                logging.info("DELETING...")
+                bot.delete_message(chat_id=update.message.chat.id, message_id=update.message.message_id)
